Update patch set 12
Patch Set 12:

We are seeing a new failure type in CSIT tests, related to stats. The failure happens only sometimes, but when it happens, it generates large output. The stats info in CSIT is processed by this [0] script.
Here is a good case [1], and a bad case [2] (some browsers fail to open the link due to the output size).

[0] https://github.com/FDio/csit/blob/493684d9a6d22dc86f8260fe8cc4244ba71b5e9e/resources/tools/papi/vpp_papi_provider.py
[1] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-perf-mrr-daily-master-2n-skx/1272/archives/log.html.gz#s1-s1-s1-s4-s10-t3-k2-k9-k6-k9-k1-k1-k1-k1-k14-k1-k1-k1-k1
[2] https://logs.fd.io/production/vex-yul-rot-jenkins-1/csit-vpp-perf-mrr-daily-master-2n-skx/1272/archives/log.html.gz#s1-s1-s1-s4-s10-t2-k2-k9-k6-k9-k1-k1-k1-k1-k14-k1-k1-k1-k1

Patch-set: 12
CC: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>