Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > there's also the fact that the VPP coder does not need to squirrel around in the code to find how to make an API MP safe.
> > 
> > So you want this to be an internal knob only, cause the external users can't possibly make any useful action based on the properties of the internal implementation in VPP...
> 
> as an internal knob i'd like it to automagically set the necessary 0 to 1 so that the API becomes mp-safe. Externally i'd like it to express to users what they can expect from the API. Having all expectations expressed in the API file is IMHO better for the users than them needing to understand some VPP internal workings to find out which APIs a mp-safe and which are not. i think of this method as instead of having a wiki that lists all mp-safe APIs.

I would prefer this to be declarative in the api as well, instead of just imperative in vppapgien for the same reason.  This is the same reasoning for my request to explicitly move the services information from vppapigen to the .api definition.

This can already be used without changing the parser by adding an option = "thread-safe = true" to the message.  

If this were to be added to the json definitions, it could be consumed by tests for catching missing test coverage, as in multiple workers weren't tested.

Patch-set: 1
CC: Gerrit User 1496 <1496@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>