Update patch set 1
Patch Set 1:

> > 
> > If this were to be added to the json definitions, it could be consumed by tests for catching missing test coverage, as in multiple workers weren't tested.
> 
> Thread safe is a property of the internal vpp api handler. Why should it be exposed to users? Putting it in the API/removed from the code also increases chances of it being wrong.

First, who are "users"?  I would argue that "tests" and any language bindings are also "users".
There is more information required than is available via the json files today.  I have a change that adds the docstrings to vppapigen, for an example, that you all are welcome to.

Second,the C plugin returns it to code, does it not?

By having data defined declaratively, we are all working towards keeping vpp from becoming the next supply chain attack vector, like what happened with solarwinds.  As I see it, you are arguing for less security to reduce the need for hygiene. If we use vppapigen as the maintainer of the hygiene, you can have the best of both worlds.

I'm not really invested in this change, one way or another, other than that if I can help Neale out, I'm glad to.  Just trying to get you to see other perspectives and trying to understand yours because it does not make sense to me.

Like I said, the functionality is already there thanks to the option keyword.

Patch-set: 1