Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > > Patch Set 11:
> > > 
> > > > Patch Set 11:
> > > > 
> > > > > Patch Set 11: Code-Review-1
> > > > > 
> > > > > pls explicitly set vpp_worker_count class variable to 0 (where you originally had set it to None)
> > > > 
> > > > I don't understand this demand(?)
> > > > What does THAT accomplish?
> > > > Why should it be part of THIS patch set?
> > > See my comments to your patch set #2
> > > 
> > > You had already implemented it, just with a different type. 
> > > 
> > > The proper action is to set it in the base class, since you are acting on it in that class.
> > 
> > No.
> > 
> > Classes other than VppTestClass are denizens here too.
> > 
> > Test result class was named VppTestResult - and it originally was representing a result of a VppTestClass, but it's no longer so.
> 
> vpp_worker_count is an integral (u32) property of vpp. It is always #threads -1, is it not?  The test case should add the class attribute.  What case is there for it to None or not present?
> 
> I'm curious, what non vpp tests use the VppTestResult?  I would think it would need its own custom logic.

This patch is about printing number of workers to screen, so I don't see any reason why it should modify any logic in VppTestCase if it's not required for it to work (verified == proof it's not required).

It's actually the other way around, a test case does not use a test result. Test runner uses test result. There is a VppTestRunner (again, wrong name) which always uses VppTestResult as its result class. And of course it runs all test cases (even non-vpp ones). And I can spot another error with current_test_case_info, which is not properly updated in e.g. TestUtil (which does not spawn a vpp).

Anyhow I would suggest that this discussion does not continue within context of this patch as it has nothing to do with it.

Patch-set: 11