Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Is the point to "cover for" CLI commands which don't extract single lines from input themselves?

Not exactly ...

So the cli's 'packet-generator new' CLI is bit tricky, the user can either specify a pcap file which is one line, or define a packet-stream which is typically multiline wrapped in {}s. 

The 'packet-generator new' command is looking for multiple lines and ignores carriage returns, it keeps reading until the EOF. Therefore when you use 'packet-generator new' with 'exec' it will try to interpret all the other lines in the exec file. 

This fix enable packet-generator new to work with exec, and doesn't require modify 'packet-generator new' which keeps all the unit tests happy.

Patch-set: 6