Update patch set 1
Patch Set 1:

This is a good addition.  

I have api's that are in vpe that should be moved here, like the endian tests.  

I disagree with using this for adding routes.  That issue highlights a deficiency in the api, which extends beyond tests.  I think it is better for vpp if the api support a client passing an iterator for vpp to process.

Patch-set: 1
CC: Gerrit User 1496 <1496@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>