Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Out of curiosity, why are we "api-fying" unit tests? These are supposed to test individual components as best as possible, i.e., just run by default all possible configs that must work. Also they must be runnable from cli because that's the best way to catch bugs, i.e., run, crash in gdb, poke at it.
> 
> There are many flavors of unit testing. 
> You might argue this particular approach is borderline belonging to the unittest directory. 
> It’s allowing adding small vpp side servlets. That other tests can use. Eg add 2M routes needed by csit. That’s slow over the python api

To Florin's point, this is probably be clearer as a separate test plugin for the test framework, I don't know if it's worth the effort to break out a new plugin.  I would think that the flush commands and Neal has some link manipulation commands that should be moved to something like this. The crash command would be useful here as well.

Patch-set: 1