Update patch set 1
Patch Set 1:

[snip]
> > There are many flavors of unit testing. 
> > You might argue this particular approach is borderline belonging to the unittest directory. 
> > It’s allowing adding small vpp side servlets. That other tests can use. Eg add 2M routes needed by csit. That’s slow over the python api
> 
> Given the dependency on 1) binary api client 2) binary api infra in vpp 3) fib (or some other feature), all independent and complex components, that doesn't sound like a "unit" test :-). 
> 
> Still, we could make exceptions, but are you sure this is what you're looking for? It sounds to me that you want a way to instrument a C app to do perf (msg/s) testing against vpp's binary api. What does the unittest plugin have offer to that end? Wouldn't it be enough to spawn that app from the existing test framework or csit, and then have it report back what it got (see [1])?
> 
> [1] https://git.fd.io/vpp/tree/src/vcl/test/test_vcl.py#n26

I get the practicality of this, but it raises the question as to whether high-level language bindings are first-class citizens or not.  That's a decision.  If we have to use a separate c binary, we are conceding that the language bindings are insufficient. Between writing a dedicated c binary and extending the api to support a generator/iterator, I'd rather the effort went into the latter, even if the plugin/functionality were disabled by default.

Patch-set: 1