Update patch set 1
Patch Set 1:

> backward incompatible change

On one hand, this is a bugfix, and we can argue "API contract as intended" has not changed. But clearly "API contract as guarded by a unit test" has changed. Usually, when in doubt it is better to bump the version.

> a bump of the semver

Just to be clear, which version (file, line) has to be bumped?

According to a strict API process, if this happens to affect behavior configured via production API messages, we should create new (_v2) messages and keep the old ones still causing the buggy behavior.
But, spending effort just to extend support for a buggy behavior feels wrong to me. Especially in this case, where (if I understand it correctly) VPP creates packets with wrong data inside.
Maybe we can treat the previous behavior as a security issue, arguing it is better to break downstream users, rather than allowing them to continue to rely on an insecure behavior?
In any case, a good case study for API/behavior stability discussion.

Patch-set: 1