Update patch set 1
Patch Set 1:

For this issue, I’d say bump the semver minor, commit this fix and move on. I will highlight this in 21.06 release notes.

There was already a precedent similar to this one, with the approach as per above.

The question i would know like to ask though is: how come that we had both the code *AND* the unit tests using incorrect endianness ?

Patch-set: 1
CC: Gerrit User 320 <320@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>