Update patch set 11
Patch Set 11:

Ole unicasted me.

I will copy my reply here for everyone to follow.
---

Yes. 
I will push a new patch.  Probably this evening.

No, it is not about the cosmetic issue of a title.

Either vppunittest was subclassed into a workerAwareVppUnittest by conditionally poking in the attribute, or it was extended to always be worker aware.  I prefer the latter as the 0 worker is worker aware.  By poking in an attribute, we are keeping the change encapsulated and essentially subclasing the class.  When a separate class depends on a private implementation detail.  For an external object to depend on the property, it needs to be part of the public interface and the contract between the objects.

 I refer you to patch set 2.  The class attribute I am asking for to provide the contract between objects was added and subsequently removed.

Patch-set: 11
Reviewer: Gerrit User 1496 <1496@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>