Update patch set 11
Patch Set 11:

>  I refer you to patch set 2.  The class attribute I am asking for to provide the contract between objects was added and subsequently removed.

It seems to be added in patchset 4? Or was it added and removed outside of the patch?
But anyway, it looks like the key issue is that even if we add the attribute, if some test does not inherit from the base class where the property is added then Klement cannot rely on the attribute to be there...

Patch-set: 11
Reviewer: Gerrit User 361 <361@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>