acl: change the algorithm for cleaning the sessions from purgatory
The previous scenarios used time-based purgatory cleaning mechanism.
It did appear to work in most cases, however, in some scenarios it did cause problems.
Test a new mechanism, whereby each worker maintains a monotonously incremented counter
of the number of times it is scheduled, as well as maintaining the global atomic
maximum of those.

If we put a session in purgatory record that maximum. When deleting the sessions
from purgatory, calculate the current minimum of run_count values across all threads.
if that value is higher than the stored max_run_count, then all the threads have
finished at least one run since then, so it is safe to delete this session.

Change-Id: Id8adee5505a625cdb904bcdfa97eee1d88c75eb9
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>