Update patch set 2
Patch Set 2:

Yes! It seems the segment manager does NOT close session when it goes down. I remember that it does in v20.01. I found a commit(https://gerrit.fd.io/r/c/vpp/+/26489) which change the behavior. Em..Should we distinguish between a ESTABLISHED session and a SYN_RCVD session on unlisten ? This may means we need to track sessions that have not yet been notified to the app ?

Patch-set: 2