Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Yes! It seems the segment manager does NOT close session when it goes down. I remember that it does in v20.01. I found a commit(https://gerrit.fd.io/r/c/vpp/+/26489) which change the behavior. Em..Should we distinguish between a ESTABLISHED session and a SYN_RCVD session on unlisten ? This may means we need to track sessions that have not yet been notified to the app ?

Ow, so this is an unlisten not a detach. 26489 makes sure we don't close ESTABLISHED sessions when we unlisten but now that leaves the CREATED ones in a limbo. So yes, we should scan the segment manager's segments for sessions in created state and close them when the listener is removed (something like segment_manager_del_sessions). Might also need to check in session_stream_accept_notify that the state is CREATED before trying to accept.

Patch-set: 2