Update patch set 1
Patch Set 1:

(2 comments)

Hi Yacan, 

First of all, awesome job! Going through all the layers of the stack this cleanly is by no means trivial. I suspect you didn't know about [1], but that's okay because your ASCII art looks better :-)

I think the patch is conceptually okay but I'd like to optimize it a bit to hopefully simplify code maintenance and to try to better make use of what we already have. To that end, I think the most important thing I'd like us to leverage is the fact that the session layer is close procedure agnostic. That is, it's willing to notify the transport that application did some sort of close, but afterwards it continues to accept new data and pass io events to the app until the transport notifies it of a complete close. Moreover, it will always send notifications to the app when the transport and session have been fully deleted. 

So what's important is for the transport to understand what type of close the app did. I think we have 2 options 1) extend the current transport_close to pass additional data 2) program transport to interpret a future close as a half-close, i.e., close with rx data is fine. The first could end up a bit messy while the second feels a bit unsatisfactory. But, as it turns out from your tcp changes, that's pretty much how tcp handles a half-close. There's no tcp protocol difference between a close and a half-close, instead the half-close only avoids sending a reset if the rx fifo has unread data. 

Therefore, on shutdown event in session_node, one option would be to call session_transport_attribute with "half-closed allowed" attribute and then call session_transport_close. It's not as clean as I'd like it to be (2 steps as opposed to 1) but it seems to do the job. In fact, we could do this straight from vcl, i.e., just update the attribute and send the close afterwards. What do you think? 

This is already long so more similar comments inline wrt to vcl.

Cheers,
Florin

[1] https://wiki.fd.io/view/VPP/HostStack/SessionLayerArchitecture#Session_Closing_State_Machine

Patch-set: 1
CC: Gerrit User 193 <193@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>