Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think the patch is conceptually okay but I'd like to optimize it a bit to hopefully simplify code maintenance and to try to better make use of what we already have. To that end, I think the most important thing I'd like us to leverage is the fact that the session layer is close procedure agnostic. That is, it's willing to notify the transport that application did some sort of close, but afterwards it continues to accept new data and pass io events to the app until the transport notifies it of a complete close. Moreover, it will always send notifications to the app when the transport and session have been fully deleted. 

Agree!

> Therefore, on shutdown event in session_node, one option would be to call session_transport_attribute with "half-closed allowed" attribute and then call session_transport_close. It's not as clean as I'd like it to be (2 steps as opposed to 1) but it seems to do the job. In fact, we could do this straight from vcl, i.e., just update the attribute and send the close afterwards. What do you think? 

Does the "half-closed allowed" attribute mean that the transport layer will treat the next close request as half-close? If it is like this, should we clear the attribute after completing half-close?

In fact, we tried not to add additional states, but there is a problem. That is, the CLOSED state implies that the app has confirmed that the session is closed and would never operate on it, so the vpp can free the session safely. But for the half-close scenario, vpp still needs confirmation from the app. E.g. when half-close tcp connection receive FIN from peer, the session state is transferred from APP_CLOSED to CLOSED. Then session_transport_delete_notify() would free the session's fifo without the confirmation from the app.

Patch-set: 1