Update patch set 7
Patch Set 7: Code-Review+1

It seems earlier this week there was a merge that increased frequency of crashes (to close to 100%) even for 1-core tests. This change seems to fix that, but if we want exact performance comparison I first need to find last non-crashing commit.
Anyway, the numbers are compatible with what trending was showing, so +1 for now.

Patch-set: 7
Reviewer: Gerrit User 1263 <1263@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
Label: Code-Review=+1