Update patch set 33
Patch Set 33:

(2 comments)

> Patch Set 33: Code-Review-1
> 
> I'm struggling to understand what value bring exposing things like VLAN strip knowing that we never use vlan stripping. 
> 

VLAN stripping was added because it is exposed as a startup.conf option for the DPDK plugin by setting vlan-strip-offload (on|off) in the per-device configuration.

> I think we should only expose things which are configurable in vpp. i.e. number of queues.

So I can be clear on how to proceed, do you want the offload fields removed entirely? Or just the enum definition for vlan strip? Or something else?

Patch-set: 33