Update patch set 34
Patch Set 34:

> Patch Set 34:
> 
> > Patch Set 33:
> > 
> > (2 comments)
> > 
> > > Patch Set 33: Code-Review-1
> > > 
> > > I'm struggling to understand what value bring exposing things like VLAN strip knowing that we never use vlan stripping. 
> > > 
> > 
> > VLAN stripping was added because it is exposed as a startup.conf option for the DPDK plugin by setting vlan-strip-offload (on|off) in the per-device configuration.
> 
> That should be removed. It was a workaround for one broken NIC and use of that parameter today can just make harm.

Ok, thanks for clarifying. The enum for that offload will be removed.


> > 
> > > I think we should only expose things which are configurable in vpp. i.e. number of queues.
> > 
> > So I can be clear on how to proceed, do you want the offload fields removed entirely? Or just the enum definition for vlan strip? Or something else?
> 
> I don't know what do you want to achive with this patch. For me valid use will be to ask VPP how many queues it can provide, but that makes sense only for native drivers which support dynamic interfcae creation and deletion (DPDK is not one of them). Exporting that parameter for DPDK is useless.

One main piece of information we hope to expose is the PCI address of the device associated with a software interface.

It is also useful to have access to settings related to queues, descriptors, and offloads that are configured on a device. The DPDK plugin makes tweaks to the default values of different settings based on the type of the device. E.g. the default number of RX/TX descriptors for virtio devices is different from that of other device types. When VPP or a device does not behave as a user expects, it is useful to be able to report the settings in use by a device to troubleshoot the issue. The only current way to get some of this information AFAICT is to run 'vppctl show hardware-interfaces'.


> I don't think VPP should be replacement for NIC vendor datasheet which lists all possible features nic supports.

I can remove these fields:

max_rx_queues
max_rx_desc
min_rx_desc
max_tx_queues
max_tx_desc
min_tx_desc
rx_offloads
tx_offloads

Does that address your concern? If there's something else you want removed, let me know.

Patch-set: 34
Reviewer: Gerrit User 366 <366@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>