build: auto-generate the list of targets in "make help"
Currently the list of targets in "make help" is manually maintained,
as a sequence of "echo" commands with target names supplied manually.

Create the list dynamically based on actual target names
and help text based on comments starting with two hash signs (##)
on the same line as the target name.

Change-Id: I716ccbdfb0621e1da78d86ccbc55574a3fe2c09b
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>