Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > Doesn't the inclusion of the local udp port further violate the standard?
> > 
> > The source port should be the source port of the inner header, should it not.  Is there a reason for polarizing all traffic between a pair of vteps?   
> > 
> > https://tools.ietf.org/html/draft-ietf-nvo3-vxlan-gpe-11#page-15
> 
> Paul... I'm super sympathetic on that point... but if you look at implementations of vxlan (including GPE) in the kernel, srcport and dstport are settable, so to be able to interoperate, we need to be able to set them too.
> 
> From iproute2:
> 
> ip link add DEVICE type vxlan id VNI [ dev PHYS_DEV ] [ { group | remote } IPADDR ] [ local { IPADDR | any } ] [ ttl TTL ] [ tos TOS ] [ df DF ] [ flowlabel FLOWLABEL ] [ dstport PORT ] [ srcport MIN MAX ] [ [no]learning ] [ [no]proxy ] [ [no]rsc ] [ [no]l2miss ] [ [no]l3miss ] [ [no]udpcsum ] [ [no]udp6zerocsumtx ] [ [no]udp6zerocsumrx ] [ ageing SECONDS ] [ maxaddress NUMBER ] [ [no]external ] [ gbp ] [ gpe ]

Hi Ed,

Maybe you can clear up my confusion.  This linux implementation varies the src encapsulation port, see: [0]. 

The proposed change here treats all tunnels as pt-pt tunnels, yet even the linux docs highlight the point that vxlans are point-multipoint.("Unlike most tunnels, a VXLAN is a 1 to N network, not just point to point. A VXLAN device can learn the IP address of the other endpoint either dynamically in a manner similar to a learning bridge, or make use of statically-configured forwarding entries.") [1] 

I think we are mixing up the encapsulation srcport and the listen port.  They are separate and distinct.  

If the srcport were actually a range between min/max, as in your example in iproute2, then one would be able to use the defaults and hash based on the encapsulated packet src port, and when min == max, there would only be a single src port, but the listen port is separate and distinct, which MAY, but not MUST be the same. 

[0] https://github.com/torvalds/linux/blob/c4681547bcce777daf576925a966ffa824edd09d/drivers/net/vxlan.c#L2677
[1] https://www.kernel.org/doc/html/v5.8/networking/vxlan.html

What am I missing?

Paul

Patch-set: 5