Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > On a high level changes looks fine by going through the code. What is the plan of merging it and removing the regular backend functionality? We need time(2-3months) to try it out on our SoC since we have moved away from dpdk plugin.
> 
> Hi Nitin, we have postponed moving IPsec to a plugin for now, unless someone want to pick it up again.

Thanks Benoit for the info. Putting a formal request here that we have a plan to use IPSec backend (this way or original way which is planned to be removed) that will be superior in performance over asynchronous crypto offload. So we need IPsec backend support. Thanks

Patch-set: 3