Update patch set 3
Patch Set 3:

> Thanks Benoit for the info. Putting a formal request here that we have a plan to use IPSec backend (this way or original way which is planned to be removed) that will be superior in performance over asynchronous crypto offload. So we need IPsec backend support. Thanks

Hi Nitin, not sure what you refer to with the IPsec backend support? My understanding is the following: the old DPDK IPsec backend has been deprecated and removed in Feb by commit 6c8533d4c18460029adbe0825b8dee257805fbc8 in favour of async crypto offload.
If this is an issue for you, you should start a discussion on vpp-dev <vpp-dev@lists.fd.io> instead, as this is unrelated to this specific change.

Patch-set: 3