Update patch set 3
Patch Set 3:

In the past, the standard practice was to push fixes to stable/<release> branches, then cherry-pick to master.

However, the current practice is to submit to the master branch first, then cherry-pick to active stable/<release> branches as required.

Please cherry-pick to master and if applicable to stable/2106

Patch-set: 3
CC: Gerrit User 320 <320@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>
CC: Gerrit User 10 <10@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>