TEST: memif - deduce how often is the VPP tx memif queue is drained by the other side
How often the queue is being serviced by the other side has important implications
for performance and latency, namely, all things being equal, doubling the time
between the two drain events requires doubling the queue length in order to maintain
the same performance.

Naturally there are the latency implications of this as well.

This is an experimental patch to estimate the behavior of the consumer from the
TX side of the house.

The idea is that if between the start of current and the end of previous TX device cycles
the queue length has decreased, it must have been serviced.

Since no data to send means no tx code is executed, the heuristic also
discards any measurements if the previous tx has happened more than 0.1sec ago as
"this tiny amount of load does not warrant monitoring"

If the queue was serviced between the last two TX runs, then the counter
in a bin corresponding to the recorded time since
the last detected queue servicing is incremented. Since we necessarily have
an error in detection this time roughly equal to interval between the last
two TX moments, the delta between these two is also reflected into the
"jitter" binned counter.

If the queue is full and we can not schedule packets, record the current
in-progress wait-time into a "queue full" binned counter.

Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>
Change-Id: Id717824f9511451a1439eb049ecdc36153380084
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>