TEST: memif - deduce how often is the VPP tx memif queue is drained by the other side
How often the queue is being serviced by the other side has important implications
for performance and latency, namely, all things being equal, doubling the time
between the two drain events requires doubling the queue length in order to maintain
the same performance.

Naturally there are the latency implications of this as well.

This is an experimental patch to estimate the behavior of the consumer from the
TX side of the house.

The idea is that if between the current and the previous TX device cycles
the queue length has decreased, it must have been serviced.

This assumes every drain event clears the queue entirely, because else if
the queue had N elements before the previous enqueue, and we enqueued M elements,
and drained M-1 of them, the size of the queue before the current enqueue will be
N+1, thus this heuristic will fail.

This can be solved of course by comparing the length of the queue *after* previous
enqueue with the length of the queue *before* the current enqueue.

Change-Id: Id717824f9511451a1439eb049ecdc36153380084
Signed-off-by: Andrew Yourtchenko <ayourtch@gmail.com>