Update patch set 1
Patch Set 1:

(2 comments)

Thanks for the patch! Everything looks good just for a couple of nits. Given that we're very close to release, I should also ask if you want this to be a part of 21.06 or if we should leave it for 21.10? 

On a separate note, could you push a separate patch with the scripts/steps that you're using to test the stack with packetdrill somewhere in extras/vcl-ldpreload? Ideally we should automate that and maybe run it periodically.

Patch-set: 1
CC: Gerrit User 193 <193@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>