Update patch set 2
Patch Set 2:

I have noticed the high rate of failure on the checkstyle jobs, so a couple of comments:

1) there is no need to abandon a change and create a new one, if you do a push with the same Change-Id: ... field, the new content will become the next revision of that change.

2) consequently, there should only eve be *one* field Change-Id, as that is what Gerrit uses to group the conceptually "same" commits. I have edited the existing commit message accordingly.

I can not review the SR code since I am not an expert - you can look up the maintainer from MAINTAINERS file and add them here; but I thought to make a note with this couple of hints - hopefully it will help in the future. thanks!

Patch-set: 2
CC: Gerrit User 320 <320@6d2eb258-4fe2-443e-8a38-ca81da23d4c2>