Update patch set 2
Patch Set 2:

Now the last failure looks like an issue with the new code that you are submitting. Are you able to compile it locally ? It can save time to ensure that "make test" locally passes on your machine, before getting a next change in - this way you will probably know the result earlier.

Patch-set: 2