fib: tunnel: Pin a tunnel's egress interface to its source
Type: feature

A new flag to the tunnel type to pin a tunnel's egress interface to the
same interface on which the tunnel's source address is applied.
The tunnel will track the source prefix to determine the [changing]
source interface.
The tunnel belongs to the SA, so it is now the SA, which is stacked on
the DPO contributed by the tunnel, that provides the DPO on which the
adjacencies are stacked.
the FIB provides a new service to filter the set of paths that
contribute to forwarding. this provides a new/unshared load-balance
object, so the user will not beneift from fast-convergence techniques;
use with caution.

Signed-off-by: Neale Ranns <neale@graphiant.com>
Change-Id: I3d027d85cfbcce68869570c62bf89a239233c0b2